Changed the type of init(value) from uint8_t to uint16_t.
git-svn-id: svn+ssh://subversion.munts.net/src/trunk/ARM@24593 06ee2f96-5407-11de-8bab-f3d7589a4122

+++ arduino/libraries/FaBo_PWM_PCA9685/src/FaBoPWM_PCA9685.cpp
@@ -36,7 +36,7 @@ bool FaBoPWM::begin()
 /**
  @brief Set Config
 */
-void FaBoPWM::init(uint8_t value)
+void FaBoPWM::init(uint16_t value)
 {
   writeI2c(PCA9685_ALL_LED_ON_L_REG, 0x00);
   writeI2c(PCA9685_ALL_LED_ON_H_REG, 0x00);
